+++ src/openassetio-python/tests/package/managerApi/test_managerinterface.py
@@ -21,7 +21,7 @@ Tests for the default implementations of ManagerInterface methods.
 # pylint: disable=invalid-name,redefined-outer-name
 # pylint: disable=missing-class-docstring,missing-function-docstring
 
-from unittest.mock import Mock
+from unittest.mock import Mock, call
 
 import pytest
 
@@ -187,6 +187,54 @@ class Test_ManagerInterface_finalizedEntityVersion:
         finalized_refs = manager_interface.finalizedEntityVersion(refs, Mock(), Mock())
 
         assert finalized_refs == refs
+
+
+class Test_ManagerInterface_getWithRelationship:
+    def test_method_defined_in_python(self, method_introspector):
+        assert method_introspector.is_defined_in_python(ManagerInterface.getWithRelationship)
+        assert method_introspector.is_implemented_once(ManagerInterface, "getWithRelationship")
+
+    def test_returns_empty_list_for_each_input(self, manager_interface):
+        success_callback = Mock()
+        error_callback = Mock()
+
+        for refs in (
+            [],
+            [Mock()],
+            [Mock(), Mock()],
+        ):
+            manager_interface.getWithRelationship(
+                Mock(), refs, Mock(), Mock(), success_callback, error_callback
+            )
+
+            error_callback.assert_not_called()
+            success_callback.assert_has_calls([call(i, []) for i in range(len(refs))])
+
+            success_callback.reset_mock()
+
+
+class Test_ManagerInterface_getWithRelationships:
+    def test_method_defined_in_python(self, method_introspector):
+        assert method_introspector.is_defined_in_python(ManagerInterface.getWithRelationships)
+        assert method_introspector.is_implemented_once(ManagerInterface, "getWithRelationships")
+
+    def test_returns_empty_list_for_each_input(self, manager_interface):
+        success_callback = Mock()
+        error_callback = Mock()
+
+        for rels in (
+            [],
+            [Mock()],
+            [Mock(), Mock()],
+        ):
+            manager_interface.getWithRelationships(
+                rels, Mock(), Mock(), Mock(), success_callback, error_callback
+            )
+
+            error_callback.assert_not_called()
+            success_callback.assert_has_calls([call(i, []) for i in range(len(rels))])
+
+            success_callback.reset_mock()
 
 
 class Test_ManagerInterface__createEntityReference: